Refactor
Changes to be committed:
	modified:   client/package-lock.json
	modified:   client/package.json
	modified:   client/src/components/AdminConsole.js
	modified:   client/src/components/ConsentScripts.js
	modified:   client/src/components/FollowUp.js
	modified:   client/src/components/ManageAdministrators.js
	modified:   client/src/components/ManageParticipants.js
	modified:   client/src/pages/Dashboard.js
	modified:   client/src/pages/HomePage.js
	modified:   client/src/routes.js
	modified:   client/src/services/auth.service.js
	modified:   client/src/services/data.service.js
	modified:   client/src/slices/authSlice.js
	modified:   client/src/slices/dataSlice.js
	modified:   client/src/store.js
	modified:   client/src/style.css
	modified:   docs/mia_app_dump.dmp
	modified:   docs/mia_psqk.txt
	modified:   server/.DS_Store
	deleted:    server/administration/admin.py
	deleted:    server/administration/apis.py
	deleted:    server/administration/apps.py
	deleted:    server/administration/models.py
	deleted:    server/administration/urls.py
	modified:   server/authentication/admin.py
	modified:   server/authentication/apis.py
	modified:   server/authentication/migrations/0001_initial.py
	modified:   server/authentication/models.py
	modified:   server/authentication/services.py
	modified:   server/authentication/urls.py
	deleted:    server/chatbot/__init__.py
	deleted:    server/chatbot/admin.py
	deleted:    server/chatbot/apis.py
	deleted:    server/chatbot/migrations/0001_initial.py
	deleted:    server/chatbot/migrations/__init__.py
	deleted:    server/chatbot/models.py
	deleted:    server/chatbot/tests.py
	deleted:    server/chatbot/urls.py
	new file:   server/config/fixtures/initial.json
	modified:   server/config/settings.py
	modified:   server/config/urls.py
	renamed:    server/administration/__init__.py -> server/consentbot/__init__.py
	new file:   server/consentbot/admin.py
	new file:   server/consentbot/apis.py
	renamed:    server/chatbot/apps.py -> server/consentbot/apps.py
	new file:   server/consentbot/migrations/0001_initial.py
	renamed:    server/administration/migrations/__init__.py -> server/consentbot/migrations/__init__.py
	new file:   server/consentbot/models.py
	renamed:    server/administration/tests.py -> server/consentbot/tests.py
	new file:   server/consentbot/urls.py
	new file:   server/requirements.txt
	deleted:    server/static/.DS_Store
	deleted:    server/static/css/.DS_Store
	deleted:    server/static/docs/Example_ConsentForm_UCIGREGoR.pdf
	deleted:    server/static/docs/UCI_UDN_Consent_Adult_Participant.pdf
	deleted:    server/static/images/.DS_Store
	deleted:    server/static/images/cake.png
	deleted:    server/static/images/dna_cookbook.png
	deleted:    server/static/images/dna_cookbook_recipe.png
	deleted:    server/static/images/dna_cookbook_variant.png
	deleted:    server/static/images/dna_from_parents.png
	deleted:    server/static/images/hipaa_compliant.png
	deleted:    server/static/images/person_cell_dna.png
	deleted:    server/static/images/readme/consent_chat.png
	deleted:    server/static/images/readme/scripts_admin.png
	deleted:    server/static/images/readme/user_chat_link.png
	deleted:    server/static/images/readme/users_admin.png
	deleted:    server/static/images/uci_health_logo.png
	deleted:    server/static/js/main.js
	deleted:    server/templates/admin.html
	deleted:    server/templates/admin_base.html
	deleted:    server/templates/base.html
	deleted:    server/templates/chat.html
	deleted:    server/templates/follow_up.html
	deleted:    server/templates/index.html
	deleted:    server/templates/login.html
	deleted:    server/templates/members.html
	deleted:    server/templates/script_editor.html
	deleted:    server/templates/scripts.html
	deleted:    server/templates/splash_chat.html
	deleted:    server/templates/users.html
	deleted:    server/templates/view_script.html
	new file:   server/utils/cache.py
	new file:   server/utils/db_reset.sh
	modified:   server/utils/utility_functions.py

+++ client/src/services/data.service.js
@@ -2,7 +2,7 @@
 import axios from "axios";
 import { store } from "../store"; // Redux store to access auth state
 
-const API = axios.create({ baseURL: "http://localhost:8000/api" });
+const API = axios.create({ baseURL: "http://localhost:8000/mia" });
 
 const getAuthHeaders = () => {
   const state = store.getState();
@@ -17,33 +17,33 @@ const getAuthHeaders = () => {
 
 // ✅ Fetch all users
 const getUsers = async () => {
-  const response = await API.get("/users", { headers: getAuthHeaders() });
+  const response = await API.get("auth/users/", { headers: getAuthHeaders() });
   return response.data;
 };
 
 // ✅ Create a new user
 const createUser = async (userData) => {
-  const response = await API.post("/users", userData, { headers: getAuthHeaders() });
+  const response = await API.post("auth/users/", userData, { headers: getAuthHeaders() });
   return response.data;
 };
 
 // ✅ Update an existing user
 const updateUser = async (userData) => {
-  const response = await API.put(`/users/${userData.user_id}`, userData, { headers: getAuthHeaders() });
+  const response = await API.put(`auth/users/${userData.username}/`, userData, { headers: getAuthHeaders() });
   return response.data;
 };
 
 // ✅ Delete a user
 const deleteUser = async (userId) => {
-  await API.delete(`/users/${userId}`, { headers: getAuthHeaders() });
+  await API.delete(`auth/users/${userId}`, { headers: getAuthHeaders() });
 };
 
 // ✅ Get user invite link
 const getInviteLink = async (userId) => {
-  const response = await API.get(`/users/${userId}/invite-link`, { headers: getAuthHeaders() });
+  const response = await API.get(`auth/users/${userId}/invite-link`, { headers: getAuthHeaders() });
   return response.data.link;
+  
 };
-
 // ✅ Fetch all follow-ups
 const getFollowUps = async () => {
 const response = await API.get("/follow-ups", { headers: getAuthHeaders() });
@@ -78,27 +78,6 @@ const getScripts = async () => {
   const deleteScript = async (id) => {
     await API.delete(`/scripts/${id}`, { headers: getAuthHeaders() });
   };
-
-  const getMembers = async () => {
-    const res = await API.get("/members", { headers: getAuthHeaders() });
-    return res.data;
-  };
-  
-  const addMember = async (data) => {
-    const res = await API.post("/members", data, { headers: getAuthHeaders() });
-    return res.data;
-  };
-  
-  const editMember = async (id, updates) => {
-    const res = await API.put(`/members/${id}`, updates, {
-      headers: getAuthHeaders(),
-    });
-    return res.data;
-  };
-  
-  const deleteMember = async (id) => {
-    await API.delete(`/members/${id}`, { headers: getAuthHeaders() });
-  };
   
 export const dataService = {
   getUsers,
@@ -111,9 +90,5 @@ export const dataService = {
   getScripts,
   addScript,
   editScript,
-  deleteScript,
-  getMembers,
-  addMember,
-  editMember,
-  deleteMember,
+  deleteScript
 };
